\name{testclasslinking-package}
\alias{testclasslinking-package}
\alias{testclasslinking}
\docType{package}
\title{\packageTitle{testclasslinking}}
\description{\packageDescription{testclasslinking}}
\section{Package Content}{\packageIndices{testclasslinking}}
\author{\packageAuthor{testclasslinking}}
\section{Maintainer}{\packageMaintainer{testclasslinking}}
\keyword{package}
